requirements.txt
numpy==1.26.4
pandas
Pillow
requests
tqdm
wget
nltk>=3.7
scikit_learn>=1.1.2
textaugment>=1.3.4
timm>=0.6.11
torch>=1.12.1
torchvision>=0.13.1
transformers==4.35.2
pydicom
open_clip_torch==2.23.0
matplotlib
seaborn
datasets
nltk
gdown
kagglehub
wandb